move mock app from v1 to server
