--- server/mock/tx.go
@@ -1,80 +1,0 @@
-//nolint
-package mock
-
-import (
-	"bytes"
-	"fmt"
-
-	"github.com/irisnet/irishub/modules/auth"
-	sdk "github.com/irisnet/irishub/types"
-)
-
-// An sdk.Tx which is its own sdk.Msg.
-type kvstoreTx struct {
-	key   []byte
-	value []byte
-	bytes []byte
-}
-
-var _ sdk.Tx = kvstoreTx{}
-
-func NewTx(key, value string) kvstoreTx {
-	bytes := fmt.Sprintf("%s=%s", key, value)
-	return kvstoreTx{
-		key:   []byte(key),
-		value: []byte(value),
-		bytes: []byte(bytes),
-	}
-}
-
-func (tx kvstoreTx) Route() string {
-	return "kvstore"
-}
-
-func (tx kvstoreTx) Type() string {
-	return "kvstore_tx"
-}
-
-func (tx kvstoreTx) GetMsgs() []sdk.Msg {
-	return []sdk.Msg{tx}
-}
-
-func (tx kvstoreTx) GetMemo() string {
-	return ""
-}
-
-func (tx kvstoreTx) GetSignBytes() []byte {
-	return tx.bytes
-}
-
-// Should the app be calling this? Or only handlers?
-func (tx kvstoreTx) ValidateBasic() sdk.Error {
-	return nil
-}
-
-func (tx kvstoreTx) GetSigners() []sdk.AccAddress {
-	return nil
-}
-
-func (tx kvstoreTx) GetSignatures() []auth.StdSignature {
-	return nil
-}
-
-// takes raw transaction bytes and decodes them into an sdk.Tx. An sdk.Tx has
-// all the signatures and can be used to authenticate.
-func decodeTx(txBytes []byte) (sdk.Tx, sdk.Error) {
-	var tx sdk.Tx
-
-	split := bytes.Split(txBytes, []byte("="))
-	if len(split) == 1 {
-		k := split[0]
-		tx = kvstoreTx{k, k, txBytes}
-	} else if len(split) == 2 {
-		k, v := split[0], split[1]
-		tx = kvstoreTx{k, v, txBytes}
-	} else {
-		return nil, sdk.ErrTxDecode("too many =")
-	}
-
-	return tx, nil
-}
